refactoring - structure of project
