+++ app/src/main/java/pro/devapp/newslist/logic/controllers/NavigationController.kt
@@ -1,4 +1,4 @@
-package pro.devapp.newslist.ui.main
+package pro.devapp.newslist.logic.controllers
 
 import android.os.Handler
 import androidx.fragment.app.Fragment
@@ -7,7 +7,7 @@ import pro.devapp.newslist.R
 import pro.devapp.newslist.ui.screens.newslist.ListNewsFragment
 import pro.devapp.newslist.ui.screens.splash.SplashFragment
 
-class MainPresenter(private val supportFragmentManager: FragmentManager) {
+class NavigationController(private val supportFragmentManager: FragmentManager) {
     fun openInitFragment(){
         val currentFragment = supportFragmentManager.findFragmentByTag("main")
         if(currentFragment == null){
