+++ app/src/main/java/pro/devapp/newslist/ui/screens/newslist/ListNewsFragment.kt
@@ -5,21 +5,21 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Observer
 import kotlinx.android.synthetic.main.fragment_news_list.*
 import org.koin.android.ext.android.inject
 import org.koin.core.parameter.parametersOf
 import pro.devapp.newslist.R
 import pro.devapp.newslist.logic.models.ModelItemNews
-import pro.devapp.newslist.ui.main.MainPresenter
+import pro.devapp.newslist.logic.controllers.NavigationController
 import pro.devapp.newslist.ui.screens.newslist.customviews.NewsList
 import pro.devapp.newslist.ui.main.NavigationFragment
 import pro.devapp.newslist.ui.screens.newsdetail.ViewNewsFragment
+import pro.devapp.newslist.util.observeNonNull
 
 class ListNewsFragment : NavigationFragment() {
 
     private val listPresenter : ListPresenter by inject()
-    private val mainPresenter : MainPresenter by inject{ parametersOf( (activity as AppCompatActivity).supportFragmentManager) }
+    private val navigationController : NavigationController by inject{ parametersOf( (activity as AppCompatActivity).supportFragmentManager) }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -33,18 +33,18 @@ class ListNewsFragment : NavigationFragment() {
         super.onActivityCreated(savedInstanceState)
         setTitle(getString(R.string.app_name))
 
-        listPresenter.pagedList.observe(viewLifecycleOwner, Observer {
+        listPresenter.pagedList.observeNonNull(viewLifecycleOwner){
             newsList.submitList(it)
-        })
+        }
 
-        listPresenter.getErrorMessage().observe(viewLifecycleOwner, Observer {
+        listPresenter.getErrorMessage().observeNonNull(viewLifecycleOwner){
             newsList.setError(it)
-        })
+        }
 
         newsList.listener = object : NewsList.ActionListener{
             override fun onItemClick(item: ModelItemNews?) {
                 item?.let {
-                    mainPresenter.addFragment(
+                    navigationController.addFragment(
                         ViewNewsFragment.newInstance(
                             it.id
                         )
