+++ app/src/main/java/pro/devapp/newslist/util/LiveDataExtension.kt
@@ -1,0 +1,13 @@
+package pro.devapp.newslist.util
+
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.Observer
+
+fun <T> LiveData<T>.observe(owner: LifecycleOwner, f: (T?) -> Unit) {
+    this.observe(owner, Observer<T> { t -> f(t) })
+}
+
+fun <T> LiveData<T>.observeNonNull(owner: LifecycleOwner, f: (T) -> Unit) {
+    this.observe(owner, Observer<T> { t -> t?.let(f) })
+}
